整理のために client.go をおく
